# apps/docs-service/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Academic Management Microservices API
  description: >-
    This is the unified API documentation for the Academic Management System, built using a production-grade microservices architecture. 
    All services are orchestrated via the Kong API Gateway, providing a single entry point for all client requests.

    The system is composed of several specialized services:
    - **User Service**: Handles identity management for students, teachers, and staff.
    - **Academic Service**: Manages course catalogs, semesters, scheduling, and notifications.
    - **Registry Service**: Governs financial transactions, fee management, and academic reporting.
    - **Messenger Service**: Facilitates real-time communication between system participants.

    Auth0 is utilized for secure, centralized authentication and authorization across the entire ecosystem.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api/v1
    description: Kong API Gateway (Local)
  - url: https://api.example.com/api/v1
    description: Production Gateway
...
security:
  - bearerAuth: []
paths:
...
  /user/student/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Students]
      summary: Get student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Students]
      summary: Update student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdateDTO'
            example:
              fullName: 'John Doe'
              enrolmentDate: '2024-01-10T00:00:00.000Z'
              status: 'GRADUATED'
              session: '2024-2025'
              graduationYear: 2028
              currentSemesterId: '5e4b9e56-1ec3-4a1a-a1f8-81e7e9a8b2fa'
      responses:
        '202':
          description: Student's updated data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Students]
      summary: Delete user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /user/teacher/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Teachers]
      summary: Get single teacher
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Teachers]
      summary: Update teacher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherUpdateDTO'
            example:
              status: graduated
      responses:
        '202':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Teachers]
      summary: Delete teacher
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /user/teacher/{teacherId}/attendance/{id}:
    parameters:
      - name: teacherId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [TeachersAttendance]
      summary: Get single teacher
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherAttendance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [TeachersAttendance]
      summary: Update teacher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherAttendanceUpdateDto'
      responses:
        '202':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherAttendance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [TeachersAttendance]
      summary: Delete teacher
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /user/attendance/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Attendance]
      summary: Get attendance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Attendance record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Attendance]
      summary: Update attendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceUpdateDTO'
      responses:
        '202':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Attendance]
      summary: Delete attendance
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DesignationEnum:
      type: string
      enum:
        - ADMIN
        - MANAGER
        - EMPLOYEE
        - INTERN
    # Auth Schemas
...
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        enrolmentDate:
          type: string
          format: date
        status:
          type: string
          default: 'ACTIVE'
        session:
          type: string
          default: 2021-22
        graduationYear:
          type: integer
          default: 2026
        currentSemesterId:
          type: string
          format: uuid

      required:
        - id
        - userId
        - status
        - session
        - currentSemesterId
...
    StudentUpdateDTO:
      type: object
      required:
        - fullName
        - enrolmentDate
        - currentSemesterId
      properties:
        fullName:
          type: string
        enrolmentDate:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - SUSPENDED
            - GRADUATED
            # Add all values from StudentStatus enum
          nullable: true
        session:
          type: string
          nullable: true
        graduationYear:
          type: integer
          format: int32
          nullable: true
        currentSemesterId:
          type: string
          format: uuid
          description: Current Semester Id must be valid!
    # Teacher
    Teacher:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        designation:
          allOf:
            - $ref: '#/components/schemas/DesignationEnum'
            - default: 'junior_instructor'
        joinDate:
          type: string
          format: date
        officeLocation:
          type: string
          default: COM-201
      required: [id, userId]
...
    TeacherUpdateDTO:
      type: object
      properties:
        fullName:
          type: string
          default: John Doe
        designation:
          allOf:
            - $ref: '#/components/schemas/DesignationEnum'
            - default: junior_instructor
        joinDate:
          type: string
          format: date
        officeLocation:
          type: string
          default: COM-201
    # Teacher Unavailability
...
    TeacherAttendance:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the unavailability record.
          example: 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'
        teacherId:
          type: string
          format: uuid
          description: The UUID of the teacher associated with this record.
          example: '550e8400-e29b-41d4-a716-446655440000'
        dateTime:
          type: string
          format: date-time
          description: The start timestamp of the unavailability (ISO 8601).
          example: '2023-12-01T09:00:00.000Z'
        status:
          type: string
          default: PRESENT
        recordedById:
          type: string
          format: uuid
          description: The UUID of the recoded system.
          example: '550e8400-e29b-41d4-a716-446655440000'
        notes:
          type: string
          description: The notes.
          maxLength: 255
          nullable: true
          example: 'Doctor Appointment'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the record was created.
          example: '2023-11-30T10:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the record was last updated.
          example: '2023-11-30T12:00:00.000Z'
...
    TeacherAttendanceUpdateDto:
      type: object
      properties:
        notes:
          type: string
          description: The notes.
          maxLength: 255
          nullable: true
          example: 'Doctor Appointment'
    # Semester
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        status:
          type: string
          default: PRESENT
        checkInTime:
          type: string
          default: 10:00
        checkOutTime:
          type: string
          default: 10:30
        notes:
          type: string
          default: Note goes here
        courseId:
          type: string
          format: uuid
        recordedBy:
          type: string
          format: uuid
...
    AttendanceUpdateDTO:
      type: object
      properties:
        status:
          type: string
          default: PRESENT
        checkInTime:
          type: string
          default: 10:00
        checkOutTime:
          type: string
          default: 10:30
        notes:
          type: string
          default: Everything good
    # Activities
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          example:
            code: 400
            message: Bad Request
            fields:
              - field: field1
                values:
                  - field1 cannot be empty.
                  - field1 cannot be empty.

              - field: field2
                values:
                  - field2 must be at least 8 characters long.
                  - field2 must be at least 8 characters long.
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          example:
            code: 401
            statusCode: Unauthorized
    Forbidden:
      description: Forbidden
      content:
        application/json:
          example:
            statusCode: 403
            message: Forbidden resource
            error: Forbidden
    NotFound:
      description: Not Found
      content:
        application/json:
          example:
            statusCode: 404
            message: 'Not found!'
            error: 'Not found!'
...