# apps/user-service/src/modules/identity/student/student.controller.ts
import { SetPermissions } from '@app/common/auth/decorators';
import { GrpcToHttpInterceptor } from '@app/grpc/interceptor/grpc-to-http.interceptor';
import { CacheTTL } from '@nestjs/cache-manager';
import {
  Body,
  Controller,
  Delete,
  Get,
  Param,
  Patch,
  Post,
  Query,
  UseInterceptors,
} from '@nestjs/common';
import { CreateStudentDto } from './dto/create-student.dto';
import { PaginationDto } from './dto/pagination.dto';
import { UpdateStudentDto } from './dto/update-student.dto';
import { StudentService } from './student.service';
@UseInterceptors(GrpcToHttpInterceptor)
@Controller('student')
@CacheTTL(1000 * 60 * 60 * 24 * 14) // 14 days
export class StudentController {
  constructor(private readonly studentService: StudentService) {}

  @SetPermissions('students:create')
  @Post()
  create(@Body() createStudentDto: CreateStudentDto) {
    return this.studentService.create(createStudentDto);
  }

  @SetPermissions('students:read')
  @Get()
  findAll(@Query() pagination: PaginationDto) {
    return this.studentService.findAll({
      limit: pagination.limit || 10,
      offset: pagination.offset || 0,
    });
  }

  @SetPermissions('students:read')
  @Get(':id')
  findOne(@Param('id') id: string) {
    return this.studentService.findOne(id);
  }

  @SetPermissions('students:update')
  @Patch(':id')
  update(@Param('id') id: string, @Body() updateStudentDto: UpdateStudentDto) {
    return this.studentService.update(id, updateStudentDto);
  }

  @SetPermissions('students:delete')
  @Delete(':id')
  remove(@Param('id') id: string) {
    return this.studentService.remove(id);
  }
}